requirements.txt
django
djangorestframework
drf-spectacular
gunicorn
whitenoise[brotli]
dj-database-url
python-dotenv
PyYAML
google-generativeai
